fix: more reconnection resync
